Made inactive touches do a touchDown to remap the touch to another zone or be free
This allows the developer to unassign a touch and it will be
automatically reassigned properly, even with the touch still down.

Before unassigned touches were simply reassigned to the same zone if the
touch was still down (they locked to a zone once touching one), so now
the touch can be reused and not have to be lifted up after touching a
zone.

Unassigning touch points should not be used when using drag(), rst(),
rnt(), etc, as these methods use the touchpoints to move/resize/rotate
the zone and so should always stay assigned to the zone to funtion
correctly