Make extending Zone easier, and many other fixes
This changes to allow overriding drawImpl(), touchImpl(), etc instead of
draw() touch(), so as to act the same as the reflection methods.

Changes built-in zones to use these methods for simplicity and testing

Also the warnings that the methods dont exist are not displayed when
these are overridden now

Change some of the example sketches to use this method too.

Also fixed textZone, and changed the pressed color on ButtonZone

Changed SMTTouchManager to use map<Touch,Zone> instead of map<Long,Zone>

Moved the tint/stroke/fill restore code to endPickDraw() instead of
beginDraw(), as there are other zones that would be affected by this
now, since the draw no longer occurs right after pickDraw always on a
zone (when direct, which has been the default for a while)