warn for methods that have no matching Zone
This changes the default behaviour to give a warning when a method
exists that has the methodPrefix+ZoneName pattern, but the ZoneName does
not exist.

removes the method warnUncalled(), as it is no longer needed.

This behaviour can be disabled by setting warnUncalledMethod to false.

Make some more TouchClient methods static too