fix getCentre(), add getOrigin()
Fixes #41

Fix getCentre() to return the global coordinates of the centre of the
zone.

Add getOrigin to return the global coordinates of the origin of the zone

Also fix getX() and getY(), they were never useful before when they just
returned x and y, now they return the global coordinates.