replace press with touchUp on PieMenuZone
This makes the behaviour change to the same as other methods, where
moving off the Zone will not be reassigned to another Zone.

This allows us to use the angle alone, to be more like a marking menu.

If the user implements a press method for the PieMenuZone it will go
back to the old way of working