KeyboardZone: constructor param for font sizing
Fixes #51

The new boolean parameter exactFonts will toggle whether we use the
large font that SMT defaults to, or an exact sized font that is good
when not resized.

This changes the rasterization size  to that of the fontSize of the
font, if false, so that their is not small errors around the edges of
fonts on Keys.

This also makes some text blurry, and makes the key text ugly when
scaled, so is not a default.

Issue #15 is being worked around here.