PieMenuZone: fix transitions between submenus
This makes sure we do not treat a touchUp while over a slice that is
being removed the same as a touchUp in the middle.

It also fixes all the bugs caused by not calling transtionToSliceRoot()
when sliceRoot is changed, which occured after a touchup on a zone being
removed and on a press of a slice that has submenus.