add SMT.debug to control debug messages
Fixes #101

This replaces the use of TouchDraw.DEBUG to determine when to output
debug messages, although TouchDraw.DEBUG will continue to act the same
to draw touches with debug info.

Also now hides the messages from TouchSources unless debug is true, as
these do not give useful information unless we are trying to debug
stuff, and could confuse users.