single tree for all zones, fixed examples
Fixes # 121

This eliminates some code in SMT that was dealing with having multiple
top-level zones, and gets rid of the multiple independent trees and the
all encompassing list of Zones, instead having a special zone SMT.sketch
which is the root of the tree of all the zones currently being drawn.

This change was tested with all examples, and those that were not
working were fixed, some of them broken by previous changes.