Pre-render indirect Zones, misc fixes
This changes to pre render indirect zones into their parents, for
supporting nesting of indirect Zones, which allow rendering with proper
ordering compared to direct Zones.

Zone's Renderer is now actually used if indirect, which allows JAVA2D,
P2D graphics to be used in Zones, although only as images of these
graphics.

SMT.sketch (the root Zone) made public again, as it was a mistake that
lead it to being package visible only.