unify add()/remove() code into Zone, change clearZones() to clearChildren() in Zone
SMT will now just act as a wrapper for the methods of the SMT.sketch
root Zone.

This also adds varargs support to Zone's add() and remove() methods.

Also changes the name of clearZones() to clearChildren() in Zone which
is more descriptive, and also add clearZones() to SMT, as a wrapper for
SMT.sketch.clearChildren(), as top-level Zones are not thought of as
children even though they are children of the root Zone.